+++ RELEASE-NOTES.md
@@ -6,6 +6,7 @@
 * #44 BQModuleProviderChecker: a test helper that ensures module provider 
       is available for auto-loading
 * #46 Support classpath:/ URLs in configuration loader.
+* #47 Configuration type for abstract resource
 
 ## 0.14: